+++ yash-semantics/src/command_search.rs
@@ -40,7 +40,7 @@ use std::rc::Rc;
 use yash_env::Env;
 use yash_env::System;
 use yash_env::builtin::Builtin;
-use yash_env::builtin::Type::{Elective, Extension, Mandatory, Special, Substitutive};
+use yash_env::builtin::Type::{Special, Substitutive};
 use yash_env::function::Function;
 use yash_env::path::PathBuf;
 use yash_env::variable::Expansion;
@@ -65,16 +65,16 @@ pub enum Target {
         /// Definition of the built-in
         builtin: Builtin,
 
-        // TODO Change the type of this field to `CString`
         /// Path to the external utility that is shadowed by the substitutive
         /// built-in
         ///
-        /// The path may not necessarily be absolute. If the `$PATH` variable
+        /// This value is only used for substitutive built-ins. For other types
+        /// of built-ins, this value is always empty.
+        ///
+        /// The path may not necessarily be absolute. If the `PATH` variable
         /// contains a relative directory name and the external utility is found
         /// in that directory, the path will be relative.
-        ///
-        /// The path will be `None` if the built-in is not substitutive.
-        path: Option<CString>,
+        path: CString,
     },
 
     /// Function
@@ -84,7 +84,7 @@ pub enum Target {
     External {
         /// Path to the external utility
         ///
-        /// The path may not necessarily be absolute. If the `$PATH` variable
+        /// The path may not necessarily be absolute. If the `PATH` variable
         /// contains a relative directory name and the external utility is found
         /// in that directory, the path will be relative.
         ///
@@ -185,37 +185,34 @@ impl ClassifyEnv for Env {
 pub fn search<E: ClassifyEnv + PathEnv>(env: &mut E, name: &str) -> Option<Target> {
     let mut target = classify(env, name);
 
-    match &mut target {
-        Target::Builtin {
-            builtin,
-            path: Some(path),
-        } => {
-            assert_eq!(builtin.r#type, Substitutive);
-            // Must verify the external counterpart exists.
-            if let Some(real_path) = search_path(env, name) {
-                *path = real_path;
-            } else {
-                return None;
+    'fill_path: {
+        let path = match &mut target {
+            Target::Builtin { builtin, path } if builtin.r#type == Substitutive => {
+                // Must verify the external counterpart exists.
+                path
             }
-        }
 
-        Target::External { path } => {
-            let real_path = if name.contains('/') {
-                // Just access the given path.
-                CString::new(name).ok()
-            } else {
-                // Need to actually find it in PATH.
-                search_path(env, name)
-            };
-            if let Some(real_path) = real_path {
-                *path = real_path;
-            } else {
-                return None;
+            Target::External { path } => {
+                if name.contains('/') {
+                    // Just access the given path.
+                    *path = CString::new(name).ok()?;
+                    break 'fill_path;
+                } else {
+                    // Need to actually find it in PATH.
+                    path
+                }
             }
-        }
-        // TODO Deduplicate the above two search_path calls
-        Target::Builtin { .. } | Target::Function(_) => {
-            // Nothing to do.
+
+            Target::Builtin { .. } | Target::Function(_) => {
+                // Nothing to do.
+                break 'fill_path;
+            }
+        };
+
+        if let Some(real_path) = search_path(env, name) {
+            *path = real_path;
+        } else {
+            return None;
         }
     }
 
@@ -242,7 +239,7 @@ pub fn classify<E: ClassifyEnv>(env: &E, name: &str) -> Target {
     let builtin = env.builtin(name);
     if let Some(builtin) = builtin {
         if builtin.r#type == Special {
-            let path = None;
+            let path = CString::default();
             return Target::Builtin { builtin, path };
         }
     }
@@ -252,11 +249,7 @@ pub fn classify<E: ClassifyEnv>(env: &E, name: &str) -> Target {
     }
 
     if let Some(builtin) = builtin {
-        let path = match builtin.r#type {
-            Special => unreachable!(),
-            Mandatory | Elective | Extension => None,
-            Substitutive => Some(CString::default()),
-        };
+        let path = CString::default();
         return Target::Builtin { builtin, path };
     }
 
@@ -293,6 +286,7 @@ mod tests {
     use assert_matches::assert_matches;
     use std::collections::HashMap;
     use std::collections::HashSet;
+    use yash_env::builtin::Type::{Elective, Extension, Mandatory};
     use yash_env::function::FunctionSet;
     use yash_env::variable::Value;
     use yash_syntax::source::Location;
@@ -377,14 +371,16 @@ mod tests {
 
         assert_matches!(
             search(&mut env, "foo"),
-            Some(Target::Builtin { builtin: result, path: None }) => {
+            Some(Target::Builtin { builtin: result, path }) => {
                 assert_eq!(result.r#type, builtin.r#type);
+                assert_eq!(*path, *c"");
             }
         );
         assert_matches!(
             classify(&env, "foo"),
-            Target::Builtin { builtin: result, path: None } => {
+            Target::Builtin { builtin: result, path } => {
                 assert_eq!(result.r#type, builtin.r#type);
+                assert_eq!(*path, *c"");
             }
         );
     }
@@ -421,14 +417,16 @@ mod tests {
 
         assert_matches!(
             search(&mut env, "foo"),
-            Some(Target::Builtin { builtin: result, path: None }) => {
+            Some(Target::Builtin { builtin: result, path }) => {
                 assert_eq!(result.r#type, builtin.r#type);
+                assert_eq!(*path, *c"");
             }
         );
         assert_matches!(
             classify(&env, "foo"),
-            Target::Builtin { builtin: result, path: None } => {
+            Target::Builtin { builtin: result, path } => {
                 assert_eq!(result.r#type, builtin.r#type);
+                assert_eq!(*path, *c"");
             }
         );
     }
@@ -441,14 +439,16 @@ mod tests {
 
         assert_matches!(
             search(&mut env, "foo"),
-            Some(Target::Builtin { builtin: result, path: None }) => {
+            Some(Target::Builtin { builtin: result, path }) => {
                 assert_eq!(result.r#type, builtin.r#type);
+                assert_eq!(*path, *c"");
             }
         );
         assert_matches!(
             classify(&env, "foo"),
-            Target::Builtin { builtin: result, path: None } => {
+            Target::Builtin { builtin: result, path } => {
                 assert_eq!(result.r#type, builtin.r#type);
+                assert_eq!(*path, *c"");
             }
         );
     }
@@ -461,14 +461,16 @@ mod tests {
 
         assert_matches!(
             search(&mut env, "foo"),
-            Some(Target::Builtin { builtin: result, path: None }) => {
+            Some(Target::Builtin { builtin: result, path }) => {
                 assert_eq!(result.r#type, builtin.r#type);
+                assert_eq!(*path, *c"");
             }
         );
         assert_matches!(
             classify(&env, "foo"),
-            Target::Builtin { builtin: result, path: None } => {
+            Target::Builtin { builtin: result, path } => {
                 assert_eq!(result.r#type, builtin.r#type);
+                assert_eq!(*path, *c"");
             }
         );
     }
@@ -481,14 +483,16 @@ mod tests {
 
         assert_matches!(
             search(&mut env, "foo"),
-            Some(Target::Builtin { builtin: result, path: None }) => {
+            Some(Target::Builtin { builtin: result, path }) => {
                 assert_eq!(result.r#type, builtin.r#type);
+                assert_eq!(*path, *c"");
             }
         );
         assert_matches!(
             classify(&env, "foo"),
-            Target::Builtin { builtin: result, path: None } => {
+            Target::Builtin { builtin: result, path } => {
                 assert_eq!(result.r#type, builtin.r#type);
+                assert_eq!(*path, *c"");
             }
         );
     }
@@ -566,16 +570,16 @@ mod tests {
 
         assert_matches!(
             search(&mut env, "foo"),
-            Some(Target::Builtin { builtin: result, path: Some(path) }) => {
+            Some(Target::Builtin { builtin: result, path }) => {
                 assert_eq!(result.r#type, builtin.r#type);
-                assert_eq!(path.to_bytes(), b"/bin/foo");
+                assert_eq!(*path, *c"/bin/foo");
             }
         );
         assert_matches!(
             classify(&env, "foo"),
-            Target::Builtin { builtin: result, path: Some(path) } => {
+            Target::Builtin { builtin: result, path } => {
                 assert_eq!(result.r#type, builtin.r#type);
-                assert_eq!(path.to_bytes(), b"");
+                assert_eq!(*path, *c"");
             }
         );
     }
@@ -598,9 +602,9 @@ mod tests {
 
         assert_matches!(
             classify(&env, "foo"),
-            Target::Builtin { builtin: result, path: Some(path) } => {
+            Target::Builtin { builtin: result, path } => {
                 assert_eq!(result.r#type, builtin.r#type);
-                assert_eq!(path.to_bytes(), b"");
+                assert_eq!(*path, *c"");
             }
         );
     }
@@ -635,10 +639,10 @@ mod tests {
         env.executables.insert("/bin/foo".to_string());
 
         assert_matches!(search(&mut env, "foo"), Some(Target::External { path }) => {
-            assert_eq!(path.to_bytes(), "/bin/foo".as_bytes());
+            assert_eq!(*path, *c"/bin/foo");
         });
         assert_matches!(classify(&env, "foo"), Target::External { path } => {
-            assert_eq!(path.to_bytes(), b"");
+            assert_eq!(*path, *c"");
         });
     }
 
@@ -652,10 +656,10 @@ mod tests {
         env.builtins.insert("bar/baz", builtin);
 
         assert_matches!(search(&mut env, "bar/baz"), Some(Target::External { path }) => {
-            assert_eq!(path.to_bytes(), "bar/baz".as_bytes());
+            assert_eq!(*path, *c"bar/baz");
         });
         assert_matches!(classify(&env, "bar/baz"), Target::External { path } => {
-            assert_eq!(path.to_bytes(), b"");
+            assert_eq!(*path, *c"");
         });
     }
 
@@ -667,13 +671,13 @@ mod tests {
         env.executables.insert("/bin/foo".to_string());
 
         assert_matches!(search(&mut env, "foo"), Some(Target::External { path }) => {
-            assert_eq!(path.to_bytes(), "/usr/bin/foo".as_bytes());
+            assert_eq!(*path, *c"/usr/bin/foo");
         });
 
         env.executables.insert("/usr/local/bin/foo".to_string());
 
         assert_matches!(search(&mut env, "foo"), Some(Target::External { path }) => {
-            assert_eq!(path.to_bytes(), "/usr/local/bin/foo".as_bytes());
+            assert_eq!(*path, *c"/usr/local/bin/foo");
         });
     }
 
@@ -685,13 +689,13 @@ mod tests {
         env.executables.insert("/bin/foo".to_string());
 
         assert_matches!(search(&mut env, "foo"), Some(Target::External { path }) => {
-            assert_eq!(path.to_bytes(), "/usr/bin/foo".as_bytes());
+            assert_eq!(*path, *c"/usr/bin/foo");
         });
 
         env.executables.insert("/usr/local/bin/foo".to_string());
 
         assert_matches!(search(&mut env, "foo"), Some(Target::External { path }) => {
-            assert_eq!(path.to_bytes(), "/usr/local/bin/foo".as_bytes());
+            assert_eq!(*path, *c"/usr/local/bin/foo");
         });
     }
 
@@ -702,7 +706,7 @@ mod tests {
         env.executables.insert("foo".to_string());
 
         assert_matches!(search(&mut env, "foo"), Some(Target::External { path }) => {
-            assert_eq!(path.to_bytes(), "foo".as_bytes());
+            assert_eq!(*path, *c"foo");
         });
     }
 }